requirements.txt
Flask==2.0.1
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
Flask-CeleryExt==0.3.4
Flask-SocketIO==5.1.0
Flask-Caching==1.10.1
flask_marshmallow==0.14.0
psycopg2-binary>=2.9.1
celery>=5.1.2
redis>=3.5.3
watchgod>=0.7
click>=7.1.2
eventlet>=0.31.0
requests>=2.25.1
sqlalchemy>=1.4.20
marshmallow-sqlalchemy==0.26.1
marshmallow>=3.12.1
